src/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.3
google-cloud-compute==1.4.0
google-cloud-container==2.11.0
google-cloud-core==2.3.2
google-cloud-gke-hub==1.5.0
google-cloud-logging==3.2.1
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kubernetes==24.2.0
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.11
websocket-client==1.3.3
Werkzeug==2.2.1
zipp==3.8.1
